requirements.txt
# Core ML
torch>=2.0.0
transformers>=4.41.2
accelerate>=0.31.0
sentence-transformers>=2.7.0
scikit-learn>=1.3.0
joblib>=1.3.0

# Dataset
datasets>=2.19.0

# Clustering (Ch.5)
umap-learn>=0.5.6
hdbscan>=0.8.33
bertopic>=0.16.0

# LLM inference (Ch.6, Ch.7)
llama-cpp-python>=0.2.0
huggingface_hub>=0.23.0

# LangChain (Ch.7)
langchain>=0.1.0
langchain-community>=0.0.20

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0
pandas>=2.0.0
numpy>=1.24.0
